+++ lib/adhara.dart
@@ -12,5 +12,6 @@ export 'package:adhara/router.dart';
 export 'package:adhara/config.dart';
 export 'package:adhara/utils.dart';
 export 'package:adhara/widgets/stateful_widget.dart';
+export 'package:adhara/widgets/stateless_widget.dart';
 export 'package:adhara/app.dart';
 //export 'package:flutter/material.dart';   //TODO enable this and check if all works good